Remove leading blank line when generating run script
Some programs, such as `start-stop-deamon` expect the very
first line of the script to contain the shebang.  These programs
fail to run the script as generated by DistPlugin as it
is currently adding a leading blank line with the shebang on
the second line.

This fixes the issue by eliminating the blank line when
generating the run script.